backend/requirements.txt
flask==2.3.3
werkzeug==2.3.7
flask-cors==3.0.10
PyPDF2==3.0.1
python-dotenv==0.19.0
openai>=1.0.0,<2.0.0
gunicorn==20.1.0
pytesseract
python-magic==0.4.27
pdf2image==1.16.3
argparse 
# Note: Tesseract is a system dependency and should be installed separately, e.g., via Homebrew on macOS 